requirements.txt
psycopg2==2.9.5
aiohttp==3.8.3
aiohttp-jinja2==1.4.2
jinja2==3.0.3
asyncpg==0.27.0
pytest-aiohttp==1.0.4
pyyaml==6.0.0